+++ src/packages/client/store/index.ts
@@ -956,6 +956,7 @@ class Store
   selectMultiple(...args: Parameters<AgentActions['selectMultiple']>) { return this.agentActions.selectMultiple(...args); }
   deselectAll(...args: Parameters<AgentActions['deselectAll']>) { return this.agentActions.deselectAll(...args); }
   togglePinnedAgent(...args: Parameters<AgentActions['togglePinnedAgent']>) { return this.agentActions.togglePinnedAgent(...args); }
+  reorderPinnedAgent(...args: Parameters<AgentActions['reorderPinnedAgent']>) { return this.agentActions.reorderPinnedAgent(...args); }
   spawnAgent(...args: Parameters<AgentActions['spawnAgent']>) { return this.agentActions.spawnAgent(...args); }
   createDirectoryAndSpawn(...args: Parameters<AgentActions['createDirectoryAndSpawn']>) { return this.agentActions.createDirectoryAndSpawn(...args); }
   sendCommand(...args: Parameters<AgentActions['sendCommand']>) { return this.agentActions.sendCommand(...args); }